# Optimize relocation counting in DehydratedDataNode.GetData