@filipnavara 2024-06-08
I think there's a bug in the code. The changed profile is not a coincidence.