@filipnavara 2024-06-08
> I think there's a bug in the code. The changed profile is not a coincidence.

I re-run the code twice, checking that the results of the original and new code are identical. There were no diffs.

Re-running the profiler produces this:
```
  100%   GetData  •  1,556 ms  •  ILCompiler.DependencyAnalysis.DehydratedDataNode.GetData(NodeFactory, Boolean)
  ► 23.4%   GetValueRefOrAddDefault  •  365 ms  •  System.Collections.Generic.Dictionary`2+CollectionsMarshalHelper.GetValueRefOrAddDefault(Dictionary, TKey, out Boolean)
  ► 21.7%   NecessaryTypeSymbol  •  337 ms  •  ILCompiler.DependencyAnalysis.NodeFactory.NecessaryTypeSymbol(TypeDesc)
  ► 17.0%   FindValue  •  264 ms  •  System.Collections.Generic.Dictionary`2.FindValue(TKey)
    7.60%   IsInstanceOfInterface  •  118 ms  •  System.Runtime.CompilerServices.CastHelpers.IsInstanceOfInterface(Void*, Object)
  ► 7.59%   Sort  •  118 ms  •  System.Collections.Generic.ArraySortHelper`1.Sort(Span, Comparison)
    3.51%   Copy  •  55 ms  •  System.Array.Copy(Array, Int32, Array, Int32, Int32)
    1.69%   [Unknown]  •  26 ms
  ► 1.47%   EmitReloc  •  23 ms  •  ILCompiler.DependencyAnalysis.ObjectDataBuilder.EmitReloc(ISymbolNode, RelocType, Int32)
  ► 1.43%   MoveNext  •  22 ms  •  ILCompiler.MetadataManager+<GetDehydratableData>d__42.MoveNext()
    1.14%   AddRange  •  18 ms  •  System.Collections.Generic.Dictionary`2.AddRange(IEnumerable)
  ► 1.02%   Grow  •  16 ms  •  System.Collections.Generic.ArrayBuilder`1.Grow(Int32)
    0.86%   Encode  •  13 ms  •  Internal.Runtime.DehydratedDataCommand.Encode(Int32, Int32, Byte[])
  ► 0.44%   ToObjectData  •  6.9 ms  •  ILCompiler.DependencyAnalysis.ObjectDataBuilder.ToObjectData()
    0.39%   GetSize  •  6.0 ms  •  ILCompiler.DependencyAnalysis.Relocation.GetSize(RelocType)
    0.39%   ReadValue  •  6.0 ms  •  ILCompiler.DependencyAnalysis.Relocation.ReadValue(RelocType, Void*)
    0.28%   coreclr.dll  •  4.4 ms
    0.28%   List`1..ctor  •  4.3 ms  •  System.Collections.Generic.List`1..ctor(IEnumerable)
    0.20%   ToArray  •  3.2 ms  •  System.Collections.Generic.List`1.ToArray()
    0.13%   NodeForLinkage  •  2.0 ms  •  ILCompiler.DependencyAnalysis.EETypeNode.NodeForLinkage(NodeFactory)
    0.06%   NodeForLinkage  •  1.0 ms  •  ILCompiler.DependencyAnalysis.TentativeMethodNode.NodeForLinkage(NodeFactory)
    0.06%   get_Current  •  1.0 ms  •  ILCompiler.MetadataManager+<GetDehydratableData>d__42.get_Current()
    0.05%   ntoskrnl.exe  •  0.8 ms
    <0.01%   Dictionary`2..ctor  •  0.08 ms  •  System.Collections.Generic.Dictionary`2..ctor(Int32, IEqualityComparer)
    <0.01%   Resize  •  0.0000002 ms  •  System.Array.Resize(ref T[], Int32)
```

Must have been some profiler fluke.